Manifest parsing: single file description.

+++ src/main/scala/Bencoding.scala
@@ -3,14 +3,16 @@ package torentator
 object Bencoding {    
     sealed trait Bencode
     case class BString(value: String) extends Bencode
-    case class BInteger(value: Integer) extends Bencode
+    case class BInteger(value: Long) extends Bencode
     case class BList(value: Seq[Bencode]) extends Bencode
     case class BDictionary(value: Map[String, Bencode]) extends Bencode
 
     import scala.util.parsing.combinator._
 
     object Parser extends JavaTokenParsers {
-        def str: Parser[BString] = (wholeNumber ^^ {_.toInt}) >> {n=> ":" ~> repN(n, """.""".r)} ^^ {case x => BString(x.mkString(""))}
+        def str: Parser[BString] = (wholeNumber ^^ {_.toInt}) >> { n=> ":" ~> repN(n, """.""".r)} ^^ {
+            case x => BString(x.mkString(""))
+        }
         def int: Parser[BInteger] = ("i" ~>  wholeNumber <~ "e") ^^ {case x => BInteger(x.toInt)}
         def list: Parser[BList] = ("l" ~>  rep(expr) <~ "e") ^^ {case x => BList(x)}
         def dict: Parser[BDictionary] = ("d" ~> rep(repN(2, expr)) <~ "e") ^^ {case x =>
